feat: add camera capture (#20)
Add functionality to access the user's camera, visualize the feed on the
screen and let the user capture an image.

For now the image is saved to the users's device. This is placeholder
behavior as the image should eventually feed into the object
recognition.

How to use:

- Open the app and click the camera button at the bottom of the screen
- The browser will request permission to use the camera, approve the
request
- The camera feed should now appear.
- Make a pose and click the capture button 😁
- An snapshot taken from the camera stream should now be downloaded to
your device

Unfortunately I didn't figure out how to fake the camera for testing, so
no automated tests are included.